tensorflow2_django_api_wsgi/requirements.txt
django
djangorestframework
django-cors-headers
pillow